+++ source/pdo_sqlsrv/pdo_dbh.cpp
@@ -1302,6 +1302,7 @@ bool pdo_sqlsrv_dbh_set_attr(_Inout_ pdo_dbh_t *dbh, _In_ zend_long attr, _Inout
             break;
 
             case SQLSRV_ATTR_BATCH_ERROR_CONTINUE:
+                // Connection-level option: applies to all statements created by this DBH.
                 driver_dbh->batch_error_continue = zend_is_true(val);
                 break;
 
@@ -1531,6 +1532,7 @@ int pdo_sqlsrv_dbh_get_attr(_Inout_ pdo_dbh_t *dbh, _In_ zend_long attr, _Inout_
 
             case SQLSRV_ATTR_BATCH_ERROR_CONTINUE:
             {
+                // Connection-level option value.
                 ZVAL_BOOL(return_value, driver_dbh->batch_error_continue);
                 break;
             }