+++ source/pdo_sqlsrv/pdo_stmt.cpp
@@ -1218,7 +1218,8 @@ int pdo_sqlsrv_stmt_next_rowset( _Inout_ pdo_stmt_t *stmt )
         SQLSRV_ASSERT( driver_stmt != NULL, "pdo_sqlsrv_stmt_next_rowset: driver_data object was null" );
 
         core_sqlsrv_next_result( static_cast<sqlsrv_stmt*>( stmt->driver_data ), true,
-                     !static_cast<sqlsrv_stmt*>( stmt->driver_data )->conn->batch_error_continue );
+                 !static_cast<sqlsrv_stmt*>( stmt->driver_data )->conn->batch_error_continue,
+                 static_cast<sqlsrv_stmt*>( stmt->driver_data )->conn->batch_error_continue );
 
         if( driver_stmt->past_next_result_end == true ) {
             // Clean up remaining metadata since new_result_set() was not called